Rename file in anticipation of BIDS Dataset Group
