+++ ivadomed/loader/files_dataset_group.py
